Added logo to the PickActivity.kt. Minor code and design updates

+++ app/src/main/java/com/application/vladcelona/eximeeting/data_classes/Event.kt
@@ -5,4 +5,4 @@ import java.util.UUID
 
 data class Event(val id: UUID = UUID.randomUUID(), var title: String = "",
                  var fromDate: Date = Date(), var toDate: Date = Date(),
-                 var location: String = "", var eventStatus: Int = 0)
+                 var location: String = "", var eventStatus: Short = 0)
